Commit langue + views

+++ tp4_meteo/ViewModels/WeatherViewModel.cs
@@ -1,11 +1,11 @@
 ﻿using System.Collections.ObjectModel;
 using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Input;
+using System.Windows.Input; 
 using tp4_meteo.Data;
 using tp4_meteo.Models;
 using tp4_meteo.Services;
-using tp4_meteo.ViewModels.Commands;
+using tp4_meteo.ViewModels.Commands; 
 
 namespace tp4_meteo.ViewModels
 {
@@ -15,26 +15,34 @@ namespace tp4_meteo.ViewModels
         private readonly IMeteoService _meteoService;
         private readonly IConfigService _configService;
 
+
         public ObservableCollection<Region> Regions { get; set; } = new ObservableCollection<Region>();
         public ObservableCollection<PrevisionData> Previsions { get; set; } = new ObservableCollection<PrevisionData>();
 
+        // region 
         private Region _selectedRegion;
         public Region SelectedRegion
         {
             get => _selectedRegion;
             set
             {
-                if (Set(ref _selectedRegion, value) && _selectedRegion != null)
+                if (Set(ref _selectedRegion, value))
                 {
-                    LoadMeteoCommand.Execute(null);
+                    // changement region 
+                    if (_selectedRegion != null)
+                    {
+                        LoadMeteoCommand.Execute(null);
+                    }
                 }
             }
         }
 
+        // ajout
         private string _newNom; public string NewNom { get => _newNom; set => Set(ref _newNom, value); }
         private string _newLat; public string NewLat { get => _newLat; set => Set(ref _newLat, value); }
         private string _newLon; public string NewLon { get => _newLon; set => Set(ref _newLon, value); }
 
+        // Commandes
         public AsyncCommand LoadMeteoCommand { get; }
         public RelayCommand AddRegionCommand { get; }
         public RelayCommand DeleteRegionCommand { get; }
@@ -45,35 +53,70 @@ namespace tp4_meteo.ViewModels
             _meteoService = meteoService;
             _configService = configService;
 
+            // chargement meteo 
             LoadMeteoCommand = new AsyncCommand(async () =>
             {
                 if (SelectedRegion == null) return;
+
+                // nettoyer list 
                 Previsions.Clear();
-                var data = await _meteoService.GetPrevisionsAsync(SelectedRegion.Latitude, SelectedRegion.Longitude, _configService.ApiKey);
-                if (data != null) foreach (var d in data) Previsions.Add(d);
+
+                // cle
+                string apiKey = _configService.ApiKey;
+                if (string.IsNullOrWhiteSpace(apiKey))
+                {
+                    MessageBox.Show("Clé API manquante", "Attention");
+                    return;
+                }
+
+                // Appel Service
+                var data = await _meteoService.GetPrevisionsAsync(SelectedRegion.Latitude, SelectedRegion.Longitude, apiKey);
+
+                if (data != null)
+                {
+                    foreach (var d in data) Previsions.Add(d);
+                }
             });
 
-            AddRegionCommand = new RelayCommand(_ => {
-                if (double.TryParse(NewLat, out double lat) && double.TryParse(NewLon, out double lon) && !string.IsNullOrWhiteSpace(NewNom))
+            AddRegionCommand = new RelayCommand(_ =>
+            {
+                if (double.TryParse(NewLat.Replace(".", ","), out double lat) &&
+                    double.TryParse(NewLon.Replace(".", ","), out double lon) &&
+                    !string.IsNullOrWhiteSpace(NewNom))
                 {
                     var r = new Region { Nom = NewNom, Latitude = lat, Longitude = lon };
                     _repository.AddRegion(r);
                     Regions.Add(r);
-                    NewNom = NewLat = NewLon = "";
+                    NewNom = ""; NewLat = ""; NewLon = ""; // Reset
+                }
+                else
+                {
+                    MessageBox.Show("Coordonnées invalides.", "Erreur");
                 }
             });
 
-            DeleteRegionCommand = new RelayCommand(_ => {
+            DeleteRegionCommand = new RelayCommand(_ =>
+            {
                 if (SelectedRegion != null)
                 {
-                    _repository.DeleteRegion(SelectedRegion);
-                    Regions.Remove(SelectedRegion);
-                    SelectedRegion = null;
-                    Previsions.Clear();
+                    if (MessageBox.Show($"Supprimer {SelectedRegion.Nom} ?", "Confirmer", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+                    {
+                        _repository.DeleteRegion(SelectedRegion);
+                        Regions.Remove(SelectedRegion);
+                        SelectedRegion = null;
+                        Previsions.Clear();
+                    }
                 }
             });
 
-            foreach (var r in _repository.GetAllRegions()) Regions.Add(r);
+            ChargerRegions();
+        }
+
+        private void ChargerRegions()
+        {
+            Regions.Clear();
+            var list = _repository.GetAllRegions();
+            foreach (var item in list) Regions.Add(item);
         }
     }
 }
